fix(web): merge auth guard into proxy.ts, remove conflicting middleware.ts
Vercel raises a build error when both middleware.ts and proxy.ts exist.
Moved the admin auth cookie check into the existing proxy.ts alongside
the next-intl middleware; updated tests to import from ~/proxy.
